requirements.txt
Django==4.2.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-ckeditor==6.7.0
django-taggit==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
django-allauth==0.61.1
boto3==1.34.34
django-cors-headers==4.3.1
dj-database-url==2.1.0
setuptools>=65.5.1
wheel>=0.38.0
Pillow==10.2.0